Uses [glidesort](https://github.com/orlp/glidesort) v0.1.2 for sorting words alphabetically. Though I'm not sure this will make an appreciable performance difference until we're sorting millions of elements?