@sts10 2023-02-09
As an informal benchmark, I ran Tidy with no options on [the 4 EFF fandom word lists](https://www.eff.org/deeplinks/2018/08/dragon-con-diceware), to combine (and sort) these four lists into one. I used [hyperfine](https://github.com/sharkdp/hyperfine), with settings: 15 warm-ups runs, 50 runs.

Without glidesort: 
```
  Time (mean ± σ):     129.8 ms ±  10.2 ms    [User: 104.3 ms, System: 25.2 ms]
  Range (min … max):   116.9 ms … 158.0 ms    50 runs
```

With glidesort:
```
  Time (mean ± σ):     122.3 ms ±   4.7 ms    [User: 98.8 ms, System: 23.3 ms]
  Range (min … max):   115.2 ms … 133.4 ms    50 runs
```

So it's a little quicker!