@sts10 2023-02-09
I've now created a separate Rust project to more thoroughly benchmark `sort_by`/`sort_unstable_by`/`glidesort::sort_by` options: https://github.com/sts10/sort_bench 

In those tests, on my machine, glidesort beats the other two options by a bit.